Make message service return a dto instead of a list

+++ Backend/src/main/java/com/Backend/services/chat_service/message/dto/MessageDTOPage.java
@@ -9,5 +9,5 @@ import java.util.List;
 @Builder
 public class MessageDTOPage {
     private final List<MessageDTO> messages;
-    private final Long totalMessagesAcrossAllPage;
+    private final int totalMessagesAcrossAllPage;
 }
